Append the anonymization report by integrating automatically the results from the current anonymization exercise

This function help to automatically integrate the results from the current anonymization to the anonymization report generated during the anonymization setup. This results include the disclosure risk analysis (global risk, individual risk, SUDA (if enabled),hierarchical risk (if enabled)) and the comparison of these risk measures before and after anonymization.

saveReprtObj(
  agrisvy,
  intialObj = NULL,
  finalObj = NULL,
  unit = NULL,
  hierarchy = NULL,
  global = TRUE,
  individual = TRUE,
  suda = FALSE,
  hierarchical = FALSE,
  childName = NULL
)

Arguments

intialObj

an sdcMicro object in which no anonymization is applied yet. One can save the sdcMicro object just after its creation in a object named for example sdc_obj_initial and use it later in intialObj=sdc_obj_initial

finalObj

the sdcMicro object where all anonymization measures has been applied. One can save the last version of the sdcMicro object created for the anonymization as sdc_obj_final and use it later for in finalObj=sdc_obj_final.

unit

statistical unit of the microdata. This is used to make the analysis in the anonymization report dynamic. If the microdata is a household data, we should considered household as unit. If the microdata is a household member data, we can specify "member" as unit or "individual" as unit

hierarchy

If hierarchical=TRUE, hierarchy contain the higher hierarchical unit that should be used in the analysis of the hierarchical risk. In the anonymization of household member microdata,one should specify hierarchy="household" if the household ID is specify in the argument hhid of the sdcMicro object and that hierarchical=TRUE in the anonymization report object.

global

a logical specifying if TRUE, that the global risk analysis and comparison before and after anonymization should be included in the report.

individual

a logical specifying, if individual=TRUE, that the individual probabilistic risk and its comparison before and after anonymyzation should be included in the anonymization report.

suda

logical specifying if suda=TRUE, that the suda score and its comparison before and after anonymization should be included in the report. One has to make sure that the condition to compute SUDA2 score is satisfied, particularly that the number of categorical quasi-identifiers is higher than 2.

hierarchical

a logical specifying if (hierarchical=TRUE), that hierarchical risk and its comparison before and after anonymization should be included in the report. One has to make sure that the hierarchical risk is computed when creating the sdcMicro object, that is, the argument hhid is specified.

childName

the object name used to save component of the resulting child report in the folder containing the anonymization report.